pyproject.toml
[project]
name = "switchable_persona"
version = "0.1.0"
description = "Environment for VLM experiments with vLLM."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = []

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu130"

[project.optional-dependencies]
vlm = [
  "vllm>=0.10.0",
  "numpy",
  "torch",
  "transformers",
  "accelerate",
  "pillow",
  "safetensors",
  "huggingface-hub",
  "torchvision",
  "torchaudio",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu130"
